add lnaguage files, update compile scripts, Update pacakge.json

+++ compile.sh
@@ -2,8 +2,15 @@ cd "ink-lang"
 rm "package-lock.json"
 rm -r "node_modules"
 npm install .
-cd ..
-cd "runner"
-rm "package-lock.json"
-rm -r "node_modules"
-npm install
+npm pack --pack-destination ./pack
+
+
+if [[ $* == *--run* ]]
+then
+    cd ..
+    cd "runner"
+    rm "package-lock.json"
+    rm -r "node_modules"
+    npm install
+    npm run dev
+fi
